Create 2 different decorators for retryable sunc methods and generators.
