+++ byn/tests/test_retry.py
@@ -2,13 +2,13 @@ from unittest import mock
 
 import pytest
 
-from byn import utils
+from byn import retry
 
 
 
 def test_retryable__once():
     real_method = mock.Mock(side_effect=[ValueError, 42])
-    f = utils.retryable(real_method, ValueError)
+    f = retry.retryable(real_method, ValueError)
 
     assert f(202, '03') == 42
 
@@ -19,7 +19,7 @@ def test_retryable__once():
 def test_retryable__once_with_callback():
     callback_method = mock.Mock()
     real_method = mock.Mock(side_effect=[ValueError, 42])
-    f = utils.retryable(real_method, ValueError, callback=callback_method)
+    f = retry.retryable(real_method, ValueError, callback=callback_method)
 
     assert f(202, '03') == 42
 
@@ -30,7 +30,7 @@ def test_retryable__once_with_callback():
 
 def test_retryable__twice_error():
     real_method = mock.Mock(side_effect=[ValueError, ValueError, 42])
-    f = utils.retryable(real_method, ValueError)
+    f = retry.retryable(real_method, ValueError)
 
     with pytest.raises(ValueError):
         f(202, '03')
@@ -41,7 +41,7 @@ def test_retryable__twice_error():
 
 def test_retryable__once_but_different():
     real_method = mock.Mock(side_effect=[NotImplementedError, 42])
-    f = utils.retryable(real_method, ValueError)
+    f = retry.retryable(real_method, ValueError)
 
     with pytest.raises(NotImplementedError):
         f(202, '03')
@@ -52,7 +52,7 @@ def test_retryable__once_but_different():
 def test_retryable__twice_error_still_working():
     callback_method = mock.Mock()
     real_method = mock.Mock(side_effect=[ValueError, ValueError, 42])
-    f = utils.retryable(real_method, ValueError, retry_count=2, callback=callback_method)
+    f = retry.retryable(real_method, ValueError, retry_count=2, callback=callback_method)
 
     assert f(202, '03') == 42
 
@@ -63,7 +63,7 @@ def test_retryable__twice_error_still_working():
 
 def test_retryable__different_errors_still_working():
     real_method = mock.Mock(side_effect=[NotImplementedError, ValueError, 42])
-    f = utils.retryable(real_method, (ValueError, NotImplementedError), retry_count=2)
+    f = retry.retryable(real_method, (ValueError, NotImplementedError), retry_count=2)
 
     assert f(202, '03') == 42
